rename debug.h -> frameInfo.h
- move DebugFlags to tangram.h
