+++ core/src/gl/primitives.cpp
@@ -1,6 +1,5 @@
 #include "primitives.h"
 
-#include "debug/debug.h"
 #include "glm/mat4x4.hpp"
 #include "glm/gtc/matrix_transform.hpp"
 #include "gl/shaderProgram.h"
