+++ core/src/labels/label.cpp
@@ -2,7 +2,6 @@
 
 #include "util/geom.h"
 #include "glm/gtx/rotate_vector.hpp"
-#include "debug/debug.h"
 #include "tangram.h"
 
 namespace Tangram {
